Masked SlackService behind an interface
